extract label measurement hook
