+++ frontend/src/metabase/visualizations/echarts/graph/treemap/style.ts
@@ -2,12 +2,14 @@ import Color from "color";
 
 const NODE_LABEL_TEXT_SHADOW_COLOR = Color("black").alpha(0.5).rgb().string();
 
+export const LABEL_PADDING = 12;
+
 export const TREEMAP_CHART_STYLE = {
   nodeLabels: {
     fontWeight: 700,
     size: 12,
     textBorderWidth: 3,
-    position: [12, 12],
+    position: [LABEL_PADDING, LABEL_PADDING],
     fontFamily: "Lato, sans-serif",
     textShadowColor: NODE_LABEL_TEXT_SHADOW_COLOR,
     textShadowBlur: 4,
