+++ frontend/src/metabase/visualizations/visualizations/TreemapChart/TreemapChart.tsx
@@ -1,6 +1,5 @@
 import type { EChartsType } from "echarts/core";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import _ from "underscore";
 
 import { formatPercent } from "metabase/static-viz/lib/numbers";
 import { Box } from "metabase/ui";
@@ -13,30 +12,12 @@ import {
   getTreemapData,
 } from "metabase/visualizations/echarts/graph/treemap/model/data";
 import { getTreemapFormatters } from "metabase/visualizations/echarts/graph/treemap/model/formatters";
-import {
-  MIN_FULL_LABEL_TILE_HEIGHT,
-  MIN_LABEL_TILE_HEIGHT,
-  MIN_LABEL_TILE_WIDTH,
-  type TreemapLabelLayout,
-  type TreemapParentLabelLayout,
-  getTreemapLabelLayouts,
-  getTreemapParentLabelLayouts,
-  shouldShowParentLabels,
-} from "metabase/visualizations/echarts/graph/treemap/model/labels";
+import { shouldShowParentLabels } from "metabase/visualizations/echarts/graph/treemap/model/labels";
 import { getTreemapInlineValueIds } from "metabase/visualizations/echarts/graph/treemap/model/tooltip";
-import {
-  getNode,
-  getTreemapLayoutNodes,
-} from "metabase/visualizations/echarts/graph/treemap/model/tree";
 import type { NodeId } from "metabase/visualizations/echarts/graph/treemap/model/types";
 import { getTreemapChartOption } from "metabase/visualizations/echarts/graph/treemap/option/option";
 import { getTreemapTooltipOption } from "metabase/visualizations/echarts/graph/treemap/option/tooltip";
-import {
-  TREEMAP_CHART_STYLE,
-  getChartPadding,
-  groupHeader,
-  leafBlock,
-} from "metabase/visualizations/echarts/graph/treemap/style";
+import { getChartPadding } from "metabase/visualizations/echarts/graph/treemap/style";
 import {
   useCloseTooltipOnScroll,
   useInjectSeriesColorsClasses,
@@ -53,10 +34,7 @@ import {
   hideTreemapHoverOverlay,
   useChartEvents,
 } from "./events";
-
-// Inset from each tile edge, matching the option's `label.position` — the wrap
-// width is the rendered tile width minus this on both sides.
-const LABEL_PADDING = TREEMAP_CHART_STYLE.nodeLabels.position[0];
+import { useLabelMeasurement } from "./use-label-measurement";
 
 export const TreemapChart = ({
   rawSeries,
@@ -77,12 +55,6 @@ export const TreemapChart = ({
   const overlayRef = useRef<TreemapHoverOverlay | null>(null);
   const [viewRootId, setViewRootId] = useState<NodeId | null>(null);
   const viewRootIdRef = useRef<string | null>(null);
-  const [labelLayout, setLabelLayout] = useState<
-    Record<NodeId, TreemapLabelLayout>
-  >({});
-  const [parentLabelLayout, setParentLabelLayout] = useState<
-    Record<NodeId, TreemapParentLabelLayout>
-  >({});
   const handleViewRootChange = useCallback((id: NodeId | null) => {
     viewRootIdRef.current = id;
     setViewRootId(id);
@@ -109,8 +81,15 @@ export const TreemapChart = ({
     [chartData, settings],
   );
 
-  const showLeafValues = settings["treemap.show_leaf_values"] ?? true;
-  const showParentValues = settings["treemap.show_parent_values"] ?? true;
+  const { labelLayout, parentLabelLayout, handleLabelMeasure } =
+    useLabelMeasurement({
+      chartRef,
+      tree: chartData?.tree ?? null,
+      formatters,
+      renderingContext,
+      showLeafValues: settings["treemap.show_leaf_values"] ?? true,
+      showParentValues: settings["treemap.show_parent_values"] ?? true,
+    });
 
   const option = useMemo(() => {
     const showLeafLabels = settings["treemap.show_leaf_labels"] ?? true;
@@ -173,92 +152,6 @@ export const TreemapChart = ({
     onVisualizationClick,
   });
 
-  // Second pass of the label layout: after ECharts finishes laying out (or
-  // re-laying out on drill/resize), read each tile's rendered size and recompute
-  // which labels show and how wide they wrap. Changing only `label.show`/`width`
-  // never changes tile geometry, so the next `finished` reads identically — the
-  // deep-equal guard returns the same state reference, React bails, and the loop
-  // converges in one extra pass.
-  const handleLabelMeasure = useCallback(() => {
-    const chart = chartRef.current;
-    const tree = chartData?.tree;
-    if (!chart || !tree || !formatters) {
-      return;
-    }
-    const nodes = getTreemapLayoutNodes(chart);
-    const total = tree.reduce((sum, node) => sum + node.value, 0);
-    const formatShare = (value: number) =>
-      formatPercent(total === 0 ? 0 : value / total);
-
-    // Leaf tiles qualify for the "full" stacked block only when the value line
-    // (the widest, at the H3 font) fits the tile width, so measure it at that
-    // font. The value renders in the leaf-label font family (see the rich style
-    // in `option.ts`), so measure with the same family.
-    const nextLayout = getTreemapLabelLayouts(nodes, {
-      minTileWidth: MIN_LABEL_TILE_WIDTH,
-      minTileHeight: MIN_LABEL_TILE_HEIGHT,
-      minFullTileHeight: MIN_FULL_LABEL_TILE_HEIGHT,
-      padding: LABEL_PADDING,
-      getValueLabelWidth: (id) => {
-        // Setting off → never qualify for the "full" block (stay name-only).
-        if (!showLeafValues) {
-          return Infinity;
-        }
-        const node = getNode(id, tree);
-        if (node == null) {
-          return Infinity;
-        }
-        return renderingContext.measureText(formatters.value(node.value), {
-          size: leafBlock.value.fontSize,
-          family: TREEMAP_CHART_STYLE.nodeLabels.fontFamily,
-          weight: leafBlock.value.fontWeight,
-        });
-      },
-    });
-    setLabelLayout((prev) => (_.isEqual(prev, nextLayout) ? prev : nextLayout));
-
-    // Parent (group) header chips: a group node id is "0", "1", … — the index
-    // into the top-level tree. Measure each header's name at the chip's font
-    // style (too-narrow chips suppress the name); also measure the right-aligned
-    // value+percentage cluster (value bold + gap + percent regular) so the chip
-    // shows it only when there's room.
-    const measureHeader = (text: string, weight: number) =>
-      renderingContext.measureText(text, {
-        size: groupHeader.fontSize,
-        family: renderingContext.fontFamily,
-        weight,
-      });
-    const nextParentLayout = getTreemapParentLabelLayouts(nodes, {
-      getLabel: (id) => getNode(id, tree)?.displayName,
-      measureTextWidth: (text) => measureHeader(text, groupHeader.fontWeight),
-      getValuePercentWidth: (id) => {
-        // Setting off → header never shows the value+percentage cluster.
-        if (!showParentValues) {
-          return Infinity;
-        }
-        const node = getNode(id, tree);
-        if (node == null) {
-          return Infinity;
-        }
-        return (
-          measureHeader(formatters.value(node.value), groupHeader.fontWeight) +
-          groupHeader.valuePercentGap +
-          measureHeader(formatShare(node.value), groupHeader.percentFontWeight)
-        );
-      },
-      padding: groupHeader.paddingX,
-    });
-    setParentLabelLayout((prev) =>
-      _.isEqual(prev, nextParentLayout) ? prev : nextParentLayout,
-    );
-  }, [
-    chartData,
-    formatters,
-    renderingContext,
-    showLeafValues,
-    showParentValues,
-  ]);
-
   const allEventHandlers = useMemo(
     () => [
       ...eventHandlers,
@@ -268,12 +161,10 @@ export const TreemapChart = ({
   );
 
   // A new dataset re-renders the chart at the absolute root, so reset the
-  // tracked view root to the overview and clear the measured label map (its ids
-  // belong to the previous tree).
+  // tracked view root to the overview. (The measured label maps reset inside
+  // `useLabelMeasurement` for the same reason.)
   useEffect(() => {
     handleViewRootChange(null);
-    setLabelLayout({});
-    setParentLabelLayout({});
   }, [chartData, handleViewRootChange]);
 
   // `setOption` (run by the renderer when `option` changes) always renders at
